Fix savegame crashes on Linux
On 64bit Linux, unlike on Windows, sizeof(long) != sizeof(int).
So void idMoveableItem::Save() using WriteInt() and then
idMoveableItem::Restore() using ReadLong() doesn't fly.

Longs should be avoided in general, I eliminated them from dhewm3
for good reasons ;)

+++ neo/d3xp/Item.cpp
@@ -2120,7 +2120,7 @@ void idMoveableItem::Restore( idRestoreGame *savefile ) {
 	savefile->ReadInt( collideFrobTimer ); //  int collideFrobTimer
 
 
-	savefile->ReadLong( spawnTime ); //  int spawnTime
+	savefile->ReadInt( spawnTime ); //  int spawnTime
 	savefile->ReadBool( showItemLine ); //  bool showItemLine
 	savefile->ReadInt( itemLineHandle ); //  int itemLineHandle
 	savefile->ReadVec3( itemLineColor ); //  idVec3 itemLineColor
